refactor: inline withNormalizedHeaders in app
