+++ src/runtime/app.ts
@@ -7,6 +7,7 @@ import {
   Router,
   toNodeListener,
   fetchWithEvent,
+  splitCookiesString,
 } from "h3";
 import { createFetch, Headers } from "ofetch";
 import destr from "destr";
@@ -18,7 +19,7 @@ import { createHooks, Hookable } from "hookable";
 import type { NitroRuntimeHooks } from "./types";
 import { useRuntimeConfig } from "./config";
 import { cachedEventHandler } from "./cache";
-import { withNormalizedHeaders } from "./utils";
+import { stringifyHeaders } from "./utils";
 import { createRouteRulesHandler, getRouteRulesForPath } from "./route-rules";
 import type { $Fetch, NitroFetchRequest } from "nitropack";
 import { plugins } from "#internal/nitro/virtual/plugins";
@@ -135,3 +136,28 @@ function createNitroApp(): NitroApp {
 export const nitroApp: NitroApp = createNitroApp();
 
 export const useNitroApp = () => nitroApp;
+
+const normalizeOutgoingFetchHeaders = (headers: Headers) => {
+  const outgoingHeaders = new Headers();
+  for (const [name, header] of headers) {
+    if (name === "set-cookie") {
+      for (const cookie of splitCookiesString(stringifyHeaders(header))) {
+        outgoingHeaders.append("set-cookie", cookie);
+      }
+    } else {
+      outgoingHeaders.set(name, stringifyHeaders(header));
+    }
+  }
+  return outgoingHeaders;
+};
+
+function withNormalizedHeaders(fetch: typeof globalThis.fetch) {
+  return async (...args: Parameters<typeof fetch>) => {
+    const r = await fetch(...args);
+    return new Response(r.body, {
+      headers: normalizeOutgoingFetchHeaders(r.headers),
+      status: r.status,
+      statusText: r.statusText,
+    });
+  };
+}
